Reorganize repo and remove barrel files. Specific exports in main.ts
