+++ src/components/routerLink.ts
@@ -1,24 +1,10 @@
-import { PrefetchConfig } from '@/types/prefetch'
 import { isUrl, Url } from '@/types/url'
 import { ResolvedRoute } from '@/types/resolved'
 import { computed, defineComponent, EmitsOptions, h, InjectionKey, SetupContext, SlotsType, VNode } from 'vue'
 import { createUseRouter } from '@/compositions/useRouter'
 import { Router } from '@/types/router'
-import { RouterPushOptions } from '@/types/routerPush'
 import { createUseLink } from '@/compositions/useLink'
-
-export type ToCallback<TRouter extends Router> = (resolve: TRouter['resolve']) => ResolvedRoute | Url | undefined
-
-export type RouterLinkProps<TRouter extends Router> = RouterPushOptions & {
-  /**
-   * The url string to navigate to or a callback that returns a url string
-   */
-  to: Url | ResolvedRoute | ToCallback<TRouter>,
-  /**
-   * Determines what assets are prefetched when router-link is rendered for this route. Overrides route level prefetch.
-   */
-  prefetch?: PrefetchConfig,
-}
+import { RouterLinkProps, ToCallback } from '@/types/routerLink'
 
 type RouterLinkSlots = {
   default?: (props: {
