+++ src/compositions/useLink.ts
@@ -1,10 +1,8 @@
-import { ComputedRef, InjectionKey, MaybeRefOrGetter, Ref, computed, toValue } from 'vue'
+import { InjectionKey, MaybeRefOrGetter, computed, toValue } from 'vue'
 import { createUsePrefetching } from '@/compositions/usePrefetching'
 import { createUseRouter } from '@/compositions/useRouter'
-import { PrefetchConfig } from '@/types/prefetch'
 import { ResolvedRoute } from '@/types/resolved'
 import { RouterPushOptions } from '@/types/routerPush'
-import { RouterReplaceOptions } from '@/types/routerReplace'
 import { RouteParamsByKey } from '@/types/routeWithParams'
 import { Url, isUrl } from '@/types/url'
 import { AllPropertiesAreOptional } from '@/types/utilities'
@@ -12,53 +10,7 @@ import { createIsRoute } from '@/guards/routes'
 import { combineUrlSearchParams } from '@/utilities/urlSearchParams'
 import { isDefined } from '@/utilities/guards'
 import { Router, RouterRouteName, RouterRoutes } from '@/types/router'
-
-export type UseLink = {
-  /**
-   * A template ref to bind to the dom for automatic prefetching
-   */
-  element: Ref<HTMLElement | undefined>,
-  /**
-   * ResolvedRoute if matched. Same value as `router.find`
-   */
-  route: ComputedRef<ResolvedRoute | undefined>,
-  /**
-   * Resolved URL with params interpolated and query applied. Same value as `router.resolve`.
-   */
-  href: ComputedRef<Url | undefined>,
-  /**
-   * True if route matches current URL or is ancestor of route that matches current URL
-   */
-  isMatch: ComputedRef<boolean>,
-  /**
-   * True if route matches current URL. Route is the same as what's currently stored at `router.route`.
-   */
-  isExactMatch: ComputedRef<boolean>,
-  /**
-   * True if route matches current URL, or is a parent route that matches the parent of the current URL.
-   */
-  isActive: ComputedRef<boolean>,
-  /**
-   * True if route matches current URL exactly.
-   */
-  isExactActive: ComputedRef<boolean>,
-  /**
-   *
-   */
-  isExternal: ComputedRef<boolean>,
-  /**
-   * Convenience method for executing `router.push` with route context passed in.
-   */
-  push: (options?: RouterPushOptions) => Promise<void>,
-  /**
-   * Convenience method for executing `router.replace` with route context passed in.
-   */
-  replace: (options?: RouterReplaceOptions) => Promise<void>,
-}
-
-export type UseLinkOptions = RouterPushOptions & {
-  prefetch?: PrefetchConfig,
-}
+import { UseLink, UseLinkOptions } from '@/types/useLink'
 
 type UseLinkArgs<
   TRouter extends Router,
