--- src/errors/index.ts
@@ -1,4 +1,0 @@
-export * from './duplicateParamsError'
-export * from './metaPropertyConflict'
-export * from './routerNotInstalledError'
-export * from './useRouteInvalidError'
