+++ src/services/createResolvedRoute.ts
@@ -2,7 +2,7 @@ import { ResolvedRoute } from '@/types/resolved'
 import { Route } from '@/types/route'
 import { parseUrl } from '@/services/urlParser'
 import { assembleUrl } from '@/services/urlAssembly'
-import { RouterResolveOptions } from '@/types/RouterResolve'
+import { RouterResolveOptions } from '@/types/routerResolve'
 import { createResolvedRouteQuery } from '@/services/createResolvedRouteQuery'
 import { getStateValues } from '@/services/state'
 import { getRouteParamValues } from './paramValidation'
