+++ src/services/hooks.spec.ts
@@ -1,7 +1,7 @@
 import { expect, test, vi } from 'vitest'
 import { createResolvedRouteQuery } from '@/services/createResolvedRouteQuery'
 import { createRouterHooks } from '@/services/createRouterHooks'
-import { BeforeRouteHook } from '@/types'
+import { BeforeRouteHook } from '@/types/hooks'
 import { ResolvedRoute } from '@/types/resolved'
 import { component } from '@/utilities/testHelpers'
