--- src/types/index.ts
@@ -1,12 +1,0 @@
-export * from './hooks'
-export * from './paramTypes'
-export * from './register'
-export * from './route'
-export * from './router'
-export * from './url'
-export type { ResolvedRoute } from './resolved'
-export type { QuerySource } from './querySource'
-export type { PrefetchConfig, PrefetchStrategy } from './prefetch'
-export type { RouterPlugin } from './routerPlugin'
-export type { PropsCallbackContext } from './props'
-export type { RouterPushOptions } from './routerPush'
