+++ src/types/router.ts
@@ -1,13 +1,13 @@
 import { App, Component, InjectionKey, Ref } from 'vue'
 import { RouterHistoryMode } from '@/services/createRouterHistory'
-import { RouterRoute } from '@/services/createRouterRoute'
+import { RouterRoute } from '@/types/routerRoute'
 import { AddAfterRouteHook, AddBeforeRouteHook, WithHooks } from '@/types/hooks'
 import { PrefetchConfig } from '@/types/prefetch'
 import { ResolvedRoute } from '@/types/resolved'
 import { Route, Routes } from '@/types/route'
 import { RouterPush } from '@/types/routerPush'
 import { RouterReplace } from '@/types/routerReplace'
-import { RouterResolve, RouterResolveOptions } from '@/types/RouterResolve'
+import { RouterResolve, RouterResolveOptions } from '@/types/routerResolve'
 import { RouterReject } from './routerReject'
 import { RouterPlugin } from './routerPlugin'
 import { KeysOfUnion } from './utilities'
