+++ src/utilities/checkDuplicateNames.ts
@@ -1,5 +1,5 @@
 import { DuplicateNamesError } from '@/errors/duplicateNamesError'
-import { Routes } from '@/types'
+import { Routes } from '@/types/route'
 import { getCount } from '@/utilities/array'
 
 export function checkDuplicateNames(routes: Routes): void {
